split server content for page and popup injection
