--- src/App.css
@@ -1,24 +1,0 @@
-.app {
-    display: flex;
-    flex-direction: row;
-    height: 100%;
-    width: 100%;
-}
-
-.logo-img {
-    height: 50%;
-    margin: 0 auto
-}
-
-.text-field {
-    align-self: center;
-    width: 80%;
-}
-
-.toolbar {
-    background-color: #3B3B3B
-}
-
-.chat-toolbar {
-    background-color: #808080
-}
